Add outbound Redis implementation
This commit adds support for performing outbound HTTP requests in the SDK.
Specifically:

* it adds the Spin WIT files for outbound Redis
* it adds (a commented out) Make target to generate the bindings
(currently, the steps to re-generate the bindings are all commented in
the Makefile because we need to update the import paths to include the
directory where we generate the headers).
* it extracts a few common types in `Interop.cs`, which now contains
general `InteropString` and `Buffer` types that represent the WIT `string`
and `list<u8>` types — this means that any WIT files that uses those
two (relatively common) types can be represented in this SDK using the
new types.
* it implements the Redis GET, SET, and PUBLISH functions as exposed by
the Spin host component.
* it adds an example in `Handler.cs`, `/redis`, which uses the three
new functions added.

Signed-off-by: Radu Matei <radu.matei@fermyon.com>